Moved mutations from queries to mutations
